backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pandas>=2.2,<3.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0